\name{bmiData}
\docType{data}
\alias{bmiData}
\title{
Adolescent BMI data set (generated toy example)
}
\description{
This data set was generated to mimic data from a two-stage
randomized clinical trial studying the effect of meal replacement
shakes on adolescent obesity. It contains the following covariates
collected at the start of the first stage: "gender," "race,"
"parent_BMI," and "baseline_BMI". At the second-stage, "month4_BMI"
was collected. Variables "A1" and "A2" are the randomized treatments
at stage one and two, and "month12_BMI" is the primary outcome
collected at the end of stage two. 
}
\usage{
bmiData
}
\format{
A matrix with rows corresponding to patients.
}
\source{
Generated by Kristin A. Linn in R
}
\keyword{datasets }