\name{iqLearn-package}
\alias{iqLearn-package}
\alias{iqLearn}
\docType{package}
\title{
Interactive Q-learning
}
\description{
Estimate an optimal dynamic treatment regime using the Interactive
Q-learning (IQ-learning) algorithm when the data has been collected from a two-stage randomized trial with binary treatments.
}
\details{
\tabular{ll}{
Package: \tab iqLearn\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-01\cr
License: \tab GPL-2 \cr
}
}
\author{
Kristin A. Linn, Eric B. Laber, Leonard A. Stefanski

Maintainer: Kristin A. Linn <kalinn@ncsu.edu>
}
\references{
Linn, K. A., Laber, E. B., Stefanski, L. A. (2015) "iqLearn: Interactive Q-Learning in R", Journal of Statistical Software, 64(1), 1--25.

Laber, E. B., Linn, K. A., and Stefanski, L. A. (2014) "Interactive model building for Q-learning", Biometrika, 101(4), 831-847.
}
